resolve all linker issue. factor out LazyPoints::Sum
