+++ src/Makefile.test.include
@@ -93,7 +93,7 @@ BITCOIN_TESTS =\
   test/blsct/arith/secp256k1/secp256k1_scalar_tests.cpp \
   test/blsct/building_block/generator_deriver_tests.cpp \
   test/blsct/building_block/imp_inner_prod_arg_tests.cpp \
-  test/blsct/building_block/lazy_point_tests.cpp \
+  test/blsct/building_block/lazy_points_tests.cpp \
   test/blsct/common_tests.cpp \
   test/blsct/eip_2333/bls12_381_keygen_tests.cpp \
   test/blsct/keys_tests.cpp \
