+++ src/Makefile.test_util.include
@@ -13,8 +13,10 @@ TEST_UTIL_H = \
   blsct/arith/mcl/mcl_g1point.h \
   blsct/arith/mcl/mcl_init.h \
   blsct/arith/mcl/mcl_scalar.h \
+  blsct/arith/mcl/mcl_util.h \
   blsct/arith/secp256k1/secp256k1_point.h \
   blsct/arith/secp256k1/secp256k1_scalar.h \
+  blsct/arith/secp256k1/secp256k1_util.h \
   blsct/double_public_key.h \
   blsct/private_key.h \
   blsct/public_key.h \
